Added an "if statement" to throw an error if CartItems is empty

+++ fast-food/Controllers/HomeController.cs
@@ -194,32 +194,39 @@ namespace fast_food.Controllers
                 DateOfCreation = DateTime.Now
             };
 
-            foreach (CartItem cartItem in cart.CartItems)
+            if (cart.CartItems.Count == 0)
             {
-                OrderItem orderItem = new OrderItem()
-                {
-                    Id = Guid.NewGuid(),
-                    ItemId = cartItem.ItemId,
-                    Item = cartItem.Item,
-                    Order = order,
-                    OrderId = order.Id,
-                    Quantity = cartItem.Quantity,
-                };
-
-                order.OrderItems.Add(orderItem);
-
-                _context.OrderItem.Add(orderItem);
-
-                cart.CartItems.Remove(cartItem);
-
-                _context.CartItem.Remove(cartItem);
+                throw new ArgumentNullException(nameof(cart.CartItems));
             }
+            else
+            {
+                foreach (CartItem cartItem in cart.CartItems)
+                {
+                    OrderItem orderItem = new OrderItem()
+                    {
+                        Id = Guid.NewGuid(),
+                        ItemId = cartItem.ItemId,
+                        Item = cartItem.Item,
+                        Order = order,
+                        OrderId = order.Id,
+                        Quantity = cartItem.Quantity,
+                    };
 
-            _context.Order.Add(order);
+                    order.OrderItems.Add(orderItem);
 
-            _context.SaveChanges();
+                    _context.OrderItem.Add(orderItem);
 
-            return RedirectToAction("Index");
+                    cart.CartItems.Remove(cartItem);
+
+                    _context.CartItem.Remove(cartItem);
+                }
+
+                _context.Order.Add(order);
+
+                _context.SaveChanges();
+
+                return RedirectToAction("Index");
+            }
         }
 
         public IActionResult OrderDetails(Guid id)
